todolist/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-ipware==4.0.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.1
Pygments==2.13.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
redgreenunittest==0.1.1
sqlparse==0.4.3
uritemplate==4.1.1
django-cors-headers
